Refactor to use config js instead of in-code config
